+++ include/types/TypeUtils.hpp
@@ -14,6 +14,32 @@ namespace tnt
     template <typename... Ts>
     overload(Ts...) -> overload<Ts...>;
 
+        // thx riptutorial
+    // https://riptutorial.com/cplusplus/example/8508/recursive-lambdas
+    template <typename Fn>
+    class y_comb
+    {
+    public:
+        explicit constexpr y_comb(Fn &&fn_) noexcept(std::is_nothrow_move_constructible_v<Fn>)
+            : fn{std::move(fn_)} {}
+
+        template <typename... Args>
+        constexpr std::invoke_result_t<Fn, Args...> operator()(Args &&... args)
+            // clang-format off
+            noexcept(std::is_nothrow_invocable_v<Fn, Args...>)
+            requires std::invocable<Fn, Args...>
+        {
+            // clang-format on
+            return fn(*this, std::forward<Args>(args)...);
+        }
+
+    private:
+        Fn fn;
+    };
+
+    template <typename Fn>
+    y_comb(Fn &&) -> y_comb<std::decay_t<Fn>>;
+
     // TODO: ensure ret can multiply an int
     inline constexpr auto if_then = [](bool cond, auto const &ret) {
         return (ret * cond);
@@ -58,8 +84,13 @@ namespace tnt
     template <typename T>
     struct crtp
     {
-        inline T &base() noexcept { return static_cast<T &>(*this); }
-        inline T const &base() const noexcept { return static_cast<T const &>(*this); }
+        constexpr crtp() noexcept
+            : super{static_cast<T &>(*this)} {}
+
+        constexpr T &base() noexcept { return static_cast<T &>(*this); }
+        constexpr T const &base() const noexcept { return static_cast<T const &>(*this); }
+
+        T &super;
     };
 
     template <typename>