+++ include/utils/Assert.hpp
@@ -2,7 +2,7 @@
 #define TNT_ASSERT_UTILITIES_HPP
 
 #include "utils/Logger.hpp"
-#include "utils/TypeUtils.hpp"
+#include "types/TypeUtils.hpp"
 
 // STATIC_CHECK is used for compile time only.
 #define STATIC_CHECK(expn) typedef char __C_ASSERT__##expn[(expn) ? 1 : -1]