+++ src/doo_ecs/Bones.cpp
@@ -6,10 +6,14 @@
 
 #include "json/JsonVector.hpp"
 
+#include "utils/Benchmark.hpp"
+
 namespace tnt::doo
 {
     void bones_sys::add_object(object const &id, bone_comp const &data_)
     {
+        PROFILE_FUNCTION();
+
         if (id == length.capacity())
         {
             length.reserve(10);
@@ -24,14 +28,19 @@ namespace tnt::doo
     void bones_sys::from_json(object const &id, nlohmann::json const &j)
     {
         if (j.contains("bone"))
+        {
+            PROFILE_FUNCTION();
             add_object(id, {.joint{j["bone"]["joint"].get<tnt::Vector>()},
                             .length{j["bone"]["length"]}});
+        }
     }
 
-    void bones_sys::to_json(object const &id, nlohmann::json &j) 
+    void bones_sys::to_json(object const &id, nlohmann::json &j)
     {
         if (active.contains(id))
         {
+            PROFILE_FUNCTION();
+
             nlohmann::json &bone{j["bone"]};
             bone["joint"] = joint[id];
             bone["length"] = length[id];
@@ -40,6 +49,7 @@ namespace tnt::doo
 
     void bones_sys::remove(object const &id) noexcept
     {
+        PROFILE_FUNCTION();
         active.erase(id);
         length.erase(length.cbegin() + id);
         joint.erase(joint.cbegin() + id);
@@ -47,6 +57,7 @@ namespace tnt::doo
 
     void bones_sys::clear() noexcept
     {
+        PROFILE_FUNCTION();
         active.clear();
         length.clear();
         joint.clear();
@@ -54,6 +65,7 @@ namespace tnt::doo
 
     Vector bones_sys::bottom(object const &id) const noexcept
     {
+        PROFILE_FUNCTION();
         return {length[id] * cosf(objects.gAngle(id)), length[id] * sinf(objects.gAngle(id))};
     }
 
@@ -61,6 +73,8 @@ namespace tnt::doo
     // https://www.youtube.com/watch?v=hbgDqyy8bIw
     void bones_sys::follow(object const &id, Vector const &target) noexcept
     {
+        PROFILE_FUNCTION();
+
         Vector const &dir{target - joint[id] - objects.gPos(id)};
         objects.angle[id] += AngleOf(dir) - objects.gAngle(id);